experiments/01-crewai-react-agent/requirements.txt
# CrewAI 依赖
crewai>=0.51.0
crewai-tools>=0.8.0

# LLM 相关
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0

# 搜索和解析
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# 其他工具
python-dotenv>=1.0.0
